requirements.txt
aiogram>=3.0.0
pydantic>=1.10.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.5.0 